-Fixed bug with renison course codes and multi-versioned courses, PSYCH 101 vs PSYCH 101R eg.
TODO
-Save and load multiple calendars

+++ app/src/main/java/cah593wang/uwaterloo/cs/student/httpswww/uwcourse/MainActivity.kt
@@ -370,7 +370,7 @@ class MainActivity : AppCompatActivity() {
         val courseEditText = findViewById<EditText>(R.id.courseCodeEditTest)
         val courseString = courseEditText.text
         if (courseString.isNullOrEmpty()) return
-        val course = courseEditText.text.toString().toInt()
+        val course = courseEditText.text.toString()
         courseEditText.setText("")
 
         val depEditText = findViewById<EditText>(R.id.departmentEditText)
